feat: Consolidate store command subcommands and implement list and delete functionalities
